chore(ui): split InfoPanel into multiple files
